Fix conversion issue in layernorm fusion (#1483)
* Add implicit_conversion
